common-components/login - style email login content
